Seems to read better when being explicit.
A few times through I actually got confused when trying to
figure out if the S3_BUCKET should actually have been set to
`dotev` instead of whatever bucket you're actually using.

I realize this is sort of a no-brainer, but it'd seem like
lowering the bar to almost the floow is the best case for adoption.