add Dotenv.load!
`Dotenv.load!` is exactly the same as `Dotenv.load` except it fails fast and hard by raising
an instance of `Errno::ENOENT` if any of the filename arguments don't exist.

This changeset also includes some small refactoring to dotenv.rb since there was shared
functionality between the two methods (DRY and all that).